call_transcribe/backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
groq>=0.4.0
google-generativeai>=0.7.0
python-multipart>=0.0.9